changing dir and conflicts and stuff
